streamlit_risk_app.py

+++ streamlit_risk_app.py
@@ -80,7 +80,7 @@ def calculate_risk_summary(inputs):
 
     df = pd.DataFrame(rows)
     max_possible_score = len(inputs) * 8
-    normalized_score = round((total_score / max_possible_score) * 10) if max_possible_score > 0 else 0
+    normalized_score = int(round((total_score / max_possible_score) * 10)) if max_possible_score > 0 else 0
 
     # Risk Clustering Bonus
     categories = [r.category for r in inputs]
